Now debug all this

+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: osericol <osericol@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 17:37:13 by osericol          #+#    #+#             */
-/*   Updated: 2023/04/05 17:13:50 by osericol         ###   ########.fr       */
+/*   Updated: 2023/04/06 15:47:06 by osericol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (!s)
 		return (NULL);
-	if (start > ft_strlen(s))
+	if ((int)start > ft_strlen(s))
 		return (ft_strdup(""));
 	str = (char *)malloc(sizeof(char) * (len + 1));
 	if (!str)
